code-example (config): fix reading
