+++ src/lib/configs/code-example/code-example.ts
@@ -1,3 +1,36 @@
-import { readFileSync } from "fs";
+export const codeExample = `
+package main
 
-export const codeExample = readFileSync(`${__dirname}/code-example.go`, "utf8");
+import (
+    "fmt"
+
+    "github.com/Goscord/goscord"
+    "github.com/Goscord/goscord/discord"
+    "github.com/Goscord/goscord/gateway"
+)
+
+var client *gateway.Session
+
+func main() {
+    fmt.Println("Starting...")
+
+    client := goscord.New(&gateway.Options{ 
+        Token: "token", 
+        Intents: gateway.IntentGuildMessages,
+    })
+
+    client.On("ready", func() {
+        fmt.Println("Logged in as " + client.Me().Tag())
+    })
+
+    client.On("message", func(msg *discord.Message) {
+        if msg.Content == "ping" {
+            client.Channel.SendMessage(msg.ChannelId, "Pong ! 🏓")
+        }
+    })
+
+    client.Login()
+
+    select {}
+}
+`.trim();
